pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smythe"
version = "0.0.1"
description = "Task-based personalized agent swarms with dynamic execution topology"
readme = "Readme.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Pete Hottelet" },
]
keywords = ["agents", "swarm", "llm", "dag", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
openclaw = ["openclaw-sdk>=0.1"]
all = ["smythe[anthropic]", "smythe[openai]", "smythe[openclaw]"]
dev = ["pytest>=8.0", "pytest-asyncio>=1.0", "ruff>=0.8"]

[project.urls]
Repository = "https://github.com/petehottelet/smythe"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
